--- assignment-3/src/App.types.ts
@@ -1,13 +1,0 @@
-export interface Book {
-  name: string
-  author: string
-  topic: string
-}
-
-export interface TableDataType {
-  key: React.Key
-  name: string
-  author: string
-  topic: string
-  action: 'Delete'
-}
