+++ assignment-3/src/pages/Homepage/Homepage.types.ts
@@ -1,0 +1,13 @@
+export interface Book {
+  name: string;
+  author: string;
+  topic: string;
+}
+
+export interface TableDataType {
+  key: React.Key;
+  name: string;
+  author: string;
+  topic: string;
+  action: 'Delete';
+}
